+++ torii/hdl/_unused.py
@@ -2,19 +2,30 @@
 
 import sys
 from sys      import _getframe
-from types    import TracebackType
+from types         import TracebackType
 from typing   import Any, TypedDict, TypeVar
 from warnings import warn_explicit
 
+from ..       import diagnostics
 from ..util   import get_linter_option
 
 __all__ = (
 	'MustUse',
-	'UnusedMustUse',
 )
 
-class UnusedMustUse(Warning):
-	pass
+
+def __dir__() -> list[str]:
+	return list({*__all__, 'UnusedMustUse'})
+
+def __getattr__(name: str):
+	if name == 'UnusedMustUse':
+		from warnings import warn
+		warn(
+			f'The import of {name} from {__name__} has been deprecated and moved '
+			f'to torii.diagnostics.MustUseWarning', DeprecationWarning, stacklevel = 2
+		)
+		return diagnostics.MustUseWarning
+	raise AttributeError(f'Module {__name__!r} has no attribute {name!r}')
 
 class _MustUseCtx(TypedDict):
 	filename: str
@@ -25,7 +36,7 @@ T = TypeVar('T')
 
 class MustUse:
 	_MustUse__silence: bool = False
-	_MustUse__warning: type[Warning] = UnusedMustUse
+	_MustUse__warning: type[Warning] = diagnostics.MustUseWarning
 	_MustUse__used: bool
 	_MustUse__context: _MustUseCtx