+++ torii/hdl/ast.py
@@ -17,11 +17,12 @@ from typing            import (
 	TYPE_CHECKING, Generic, Literal, NoReturn, ParamSpec, SupportsIndex, TypeAlias, TypeVar, TypeVarTuple,
 )
 
+from ..                import diagnostics
 from .._typing         import SrcLoc, SwitchCaseT
 from ..util            import flatten, tracer, union
 from ..util.decorators import final
 from ..util.units      import bits_for
-from ._unused          import MustUse, UnusedMustUse
+from ._unused          import MustUse
 
 __all__ = (
 	'AnyConst',
@@ -66,6 +67,20 @@ __all__ = (
 	'ValueLike',
 	'ValueSet',
 )
+
+
+def __dir__() -> list[str]:
+	return list({*__all__, 'UnusedProperty'})
+
+def __getattr__(name: str):
+	if name == 'UnusedProperty':
+		from warnings import warn
+		warn(
+			f'The import of {name} from {__name__} has been deprecated and moved '
+			f'to torii.diagnostics.{name}', DeprecationWarning, stacklevel = 2
+		)
+		return diagnostics.UnusedProperty
+	raise AttributeError(f'Module {__name__!r} has no attribute {name!r}')
 
 T = TypeVar('T')
 U = TypeVar('U')
@@ -2137,12 +2152,9 @@ class Assign(Statement):
 	def __repr__(self) -> str:
 		return f'(eq {self.lhs!r} {self.rhs!r})'
 
-class UnusedProperty(UnusedMustUse):
-	pass
-
 @final
 class Property(Statement, MustUse):
-	_MustUse__warning = UnusedProperty
+	_MustUse__warning = diagnostics.UnusedProperty
 
 	class Kind(Enum):
 		Assert = 'assert'