+++ torii/hdl/xfrm.py
@@ -5,12 +5,12 @@ from collections     import OrderedDict
 from collections.abc import Iterable
 from copy            import copy
 
+from ..diagnostics   import DomainError
 from ..util          import flatten, tracer
 from .ast            import (
 	AnyValue, ArrayProxy, Assign, Cat, ClockSignal, Const, Initial, Mux, Operator, Part, Property, ResetSignal, Sample,
 	Signal, SignalDict, SignalSet, Slice, Statement, Switch, Value, ValueDict, _StatementList,
 )
-from .cd             import DomainError
 from .ir             import Elaboratable, Fragment, Instance
 from .mem            import MemoryInstance