+++ torii/tools/__init__.py
@@ -3,6 +3,7 @@
 from os            import environ
 from shutil        import which
 
+from ..            import diagnostics
 from ..util.string import tool_env_var
 
 __doc__ = '''\
@@ -17,11 +18,20 @@ __doc__ = '''\
 __all__ = (
 	'has_tool',
 	'require_tool',
-	'ToolNotFound',
 )
 
-class ToolNotFound(Exception):
-	pass
+def __dir__() -> list[str]:
+	return list({*__all__, 'ToolNotFound'})
+
+def __getattr__(name: str):
+	if name == 'ToolNotFound':
+		from warnings import warn
+		warn(
+			f'The import of {name} from {__name__} has been deprecated and moved '
+			f'to torii.diagnostics.{name}', DeprecationWarning, stacklevel = 2
+		)
+		return diagnostics.ToolNotFound
+	raise AttributeError(f'Module {__name__!r} has no attribute {name!r}')
 
 def _get_tool(name: str) -> str:
 	'''
@@ -83,12 +93,12 @@ def require_tool(name: str) -> str:
 	path = _get_tool(name)
 	if which(path) is None:
 		if env_var in environ:
-			raise ToolNotFound(
+			raise diagnostics.ToolNotFound(
 				f'Could not find required tool {name} in {path} as specified via '
 				f'the {env_var} environment variable'
 			)
 		else:
-			raise ToolNotFound(
+			raise diagnostics.ToolNotFound(
 				f'Could not find required tool {name} in PATH. Place '
 				'it directly in PATH or specify path explicitly '
 				f'via the {env_var} environment variable'